backend/requirements.txt
bugsnag==3.4.2
Django==2.0.6
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.4.0
factory-boy==2.10.0
psycopg2-binary==2.7.5
uWSGI==2.0.17
